src/seebot/requirements.txt
PySide6==6.6.0
selenium==4.15.2
Requests==2.31.0
pycryptodome